refactor(api): extract backup/restore handlers from system.rs (#3749 5/N)
routes/system.rs continues to span many unrelated domains. Extract the
backup/restore sub-domain into a sibling module routes/backup.rs,
following the pattern set by #4452 (pairing), #4454 (agent_templates),
#4455 (tools_sessions), and #4458 (hooks_commands).

Move create_backup, list_backups, delete_backup, restore_backup, the
BackupManifest struct, and the read_backup_manifest /
is_invalid_backup_filename / find_backup_path helpers. system::router()
now mounts /backup, /backups, /backups/{filename}, and /restore via
.merge(routes::backup::router()) — public route paths are unchanged.

routes/mod.rs gets pub mod backup + pub use backup::* so existing
utoipa path bindings in openapi.rs (routes::create_backup, …) continue
to resolve. hostname_string() stays in system.rs and is called via
super::system::hostname_string() from the new module.

system.rs goes from 5939 -> 5381 lines (-558); new backup.rs is 577 lines.

Refs #3749 (5 of N sub-domain extractions).
