+++ crates/librefang-api/src/routes/mod.rs
@@ -17,6 +17,7 @@ pub mod agents;
 pub mod audit;
 pub mod authz;
 pub mod auto_dream;
+pub mod backup;
 pub mod budget;
 pub mod channels;
 pub mod config;
@@ -49,6 +50,7 @@ pub use agents::*;
 pub use audit::*;
 pub use authz::*;
 pub use auto_dream::*;
+pub use backup::*;
 pub use budget::*;
 pub use channels::*;
 pub use config::*;
